Added context for 3d stuff and added example usage in navbar
