+++ apps/client/components/Navbar.tsx
@@ -1,23 +1,24 @@
-import React from "react";
+import React, { useContext } from "react";
 import NavButton from "./NavButton";
+import { RendererContext } from "../contexts/rendererContext";
 const Navbar = () => {
+  const rendererContext = useContext(RendererContext);
+  const viewer = rendererContext.viewer;
   return (
     <nav className="navbar">
       <p className="logo">SjorCRAFT</p>
       <div className="nav-buttons">
         <NavButton
           text="selectiontool"
-          onClick={() =>
-            window.dispatchEvent(new Event("activate_selectiontool"))
-          }
+          onClick={() => viewer?.inputHandler.onActivateTool("selectiontool")}
         />
         <NavButton
           text="poletool"
-          onClick={() => window.dispatchEvent(new Event("activate_poletool"))}
+          onClick={() => viewer?.inputHandler.onActivateTool("poletool")}
         />
         <NavButton
           text="bipodtool"
-          onClick={() => window.dispatchEvent(new Event("activate_bipodtool"))}
+          onClick={() => viewer?.inputHandler.onActivateTool("bipodtool")}
         />
       </div>
       <div className="profile"></div>
