+++ apps/client/contexts/rendererContext.tsx
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+import { Viewer } from "../../3d/src/viewer";
+
+export type RendererContextType = { viewer?: Viewer };
+export const RendererContext = createContext<RendererContextType>({});
